refactor: remove type_coercion in PhysicalExpr.
