+++ datafusion/core/src/physical_plan/windows/mod.rs
@@ -24,14 +24,13 @@ use crate::physical_plan::{
         cume_dist, dense_rank, lag, lead, percent_rank, rank, Literal, NthValue, Ntile,
         PhysicalSortExpr, RowNumber,
     },
-    type_coercion::coerce,
     udaf, ExecutionPlan, PhysicalExpr,
 };
 use crate::scalar::ScalarValue;
 use arrow::datatypes::Schema;
 use arrow_schema::{SchemaRef, SortOptions};
 use datafusion_expr::{
-    window_function::{signature_for_built_in, BuiltInWindowFunction, WindowFunction},
+    window_function::{BuiltInWindowFunction, WindowFunction},
     WindowFrame,
 };
 use datafusion_physical_expr::window::{
@@ -133,8 +132,7 @@ fn create_built_in_window_expr(
         BuiltInWindowFunction::PercentRank => Arc::new(percent_rank(name)),
         BuiltInWindowFunction::CumeDist => Arc::new(cume_dist(name)),
         BuiltInWindowFunction::Ntile => {
-            let coerced_args = coerce(args, input_schema, &signature_for_built_in(fun))?;
-            let n: i64 = get_scalar_value_from_args(&coerced_args, 0)?
+            let n: i64 = get_scalar_value_from_args(args, 0)?
                 .ok_or_else(|| {
                     DataFusionError::Execution(
                         "NTILE requires at least 1 argument".to_string(),
@@ -145,33 +143,26 @@ fn create_built_in_window_expr(
             Arc::new(Ntile::new(name, n))
         }
         BuiltInWindowFunction::Lag => {
-            let coerced_args = coerce(args, input_schema, &signature_for_built_in(fun))?;
-            let arg = coerced_args[0].clone();
+            let arg = args[0].clone();
             let data_type = args[0].data_type(input_schema)?;
-            let shift_offset = get_scalar_value_from_args(&coerced_args, 1)?
+            let shift_offset = get_scalar_value_from_args(args, 1)?
                 .map(|v| v.try_into())
                 .and_then(|v| v.ok());
-            let default_value = get_scalar_value_from_args(&coerced_args, 2)?;
+            let default_value = get_scalar_value_from_args(args, 2)?;
             Arc::new(lag(name, data_type, arg, shift_offset, default_value))
         }
         BuiltInWindowFunction::Lead => {
-            let coerced_args = coerce(args, input_schema, &signature_for_built_in(fun))?;
-            let arg = coerced_args[0].clone();
+            let arg = args[0].clone();
             let data_type = args[0].data_type(input_schema)?;
-            let shift_offset = get_scalar_value_from_args(&coerced_args, 1)?
+            let shift_offset = get_scalar_value_from_args(args, 1)?
                 .map(|v| v.try_into())
                 .and_then(|v| v.ok());
-            let default_value = get_scalar_value_from_args(&coerced_args, 2)?;
+            let default_value = get_scalar_value_from_args(args, 2)?;
             Arc::new(lead(name, data_type, arg, shift_offset, default_value))
         }
         BuiltInWindowFunction::NthValue => {
-            let coerced_args = coerce(args, input_schema, &signature_for_built_in(fun))?;
-            let arg = coerced_args[0].clone();
-            let n = coerced_args[1]
-                .as_any()
-                .downcast_ref::<Literal>()
-                .unwrap()
-                .value();
+            let arg = args[0].clone();
+            let n = args[1].as_any().downcast_ref::<Literal>().unwrap().value();
             let n: i64 = n
                 .clone()
                 .try_into()
@@ -181,14 +172,12 @@ fn create_built_in_window_expr(
             Arc::new(NthValue::nth(name, arg, data_type, n)?)
         }
         BuiltInWindowFunction::FirstValue => {
-            let arg =
-                coerce(args, input_schema, &signature_for_built_in(fun))?[0].clone();
+            let arg = args[0].clone();
             let data_type = args[0].data_type(input_schema)?;
             Arc::new(NthValue::first(name, arg, data_type))
         }
         BuiltInWindowFunction::LastValue => {
-            let arg =
-                coerce(args, input_schema, &signature_for_built_in(fun))?[0].clone();
+            let arg = args[0].clone();
             let data_type = args[0].data_type(input_schema)?;
             Arc::new(NthValue::last(name, arg, data_type))
         }
