+++ datafusion/physical-expr/src/lib.rs
@@ -37,7 +37,6 @@ mod sort_expr;
 pub mod string_expressions;
 pub mod struct_expressions;
 pub mod tree_node;
-pub mod type_coercion;
 pub mod udf;
 #[cfg(feature = "unicode_expressions")]
 pub mod unicode_expressions;
